backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.0
langgraph==0.6.5
langchain_community==0.3.27
langchain_anthropic==0.3.18
langchain-tavily==0.2.11
langchain_experimental==0.3.4
langchain==0.3.27
langgraph-prebuilt==0.6.4
langchain-openai==0.3.30
